--- a/Cargo.toml
+++ b/Cargo.toml
@@ -14,11 +14,16 @@ keywords = ["hashmap", "u64", "intmap"]
 serde = { version = "1.0", optional = true, default-features = false }
 
 [dev-dependencies]
-rand = "0.8.5"
+divan = "0.1.14"
 indexmap = "1.8.2"
+rand = "0.8.5"
 
 [package.metadata.docs.rs]
 features = ["serde"]
+
+[[bench]]
+name = "basic_bench"
+harness = false
 
 [workspace]
 resolver = "2"