--- a/integration_tests/Cargo.toml
+++ b/integration_tests/Cargo.toml
@@ -1,0 +1,7 @@
+[workspace]
+resolver = "2"
+members = [
+    "benchmark",
+    "random_ops",
+    "serde",
+]