This PR migrates the current benches from unstable Rust to [divan](https://crates.io/crates/divan). Nothing else. If you are happy with this, then I'll create another PR that adds more benchmarks using divan's [`args feature`](https://docs.rs/divan/0.1.14/divan/attr.bench.html#args).

You can run the benchmarks with `cargo bench`.

Before:

```
running 11 tests
test tests::u64_get_built_in                     ... bench:      80,418.25 ns/iter (+/- 1,215.98)
test tests::u64_get_indexmap                     ... bench:     104,553.57 ns/iter (+/- 716.26)
test tests::u64_get_intmap                       ... bench:      12,021.77 ns/iter (+/- 246.29)
test tests::u64_insert_built_in                  ... bench:     115,302.30 ns/iter (+/- 2,222.09)
test tests::u64_insert_built_in_without_capacity ... bench:     283,263.70 ns/iter (+/- 957.75)
test tests::u64_insert_indexmap                  ... bench:     131,009.76 ns/iter (+/- 1,660.13)
test tests::u64_insert_intmap                    ... bench:      35,768.73 ns/iter (+/- 894.52)
test tests::u64_insert_intmap_checked            ... bench:      42,419.53 ns/iter (+/- 727.90)
test tests::u64_insert_intmap_entry              ... bench:      43,489.37 ns/iter (+/- 161.42)
test tests::u64_insert_intmap_without_capacity   ... bench:     487,727.00 ns/iter (+/- 5,910.03)
test tests::u64_resize_intmap                    ... bench:      22,185.56 ns/iter (+/- 97.57)
```

After:

```
Timer precision: 10 ns
basic_bench                              fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ u64_get_built_in                      99.89 µs      │ 108.8 µs      │ 102 µs        │ 102.4 µs      │ 100     │ 100
├─ u64_get_indexmap                      106.5 µs      │ 114 µs        │ 107.3 µs      │ 107.7 µs      │ 100     │ 100
├─ u64_get_intmap                        12.2 µs       │ 37.05 µs      │ 12.31 µs      │ 12.99 µs      │ 100     │ 100
├─ u64_insert_built_in                   109.6 µs      │ 149.1 µs      │ 110.4 µs      │ 111.2 µs      │ 100     │ 100
├─ u64_insert_built_in_without_capacity  271 µs        │ 283.4 µs      │ 273.9 µs      │ 274.2 µs      │ 100     │ 100
├─ u64_insert_indexmap                   143.6 µs      │ 150.4 µs      │ 146.5 µs      │ 146.8 µs      │ 100     │ 100
├─ u64_insert_intmap                     27.17 µs      │ 151.6 µs      │ 27.77 µs      │ 29.7 µs       │ 100     │ 100
├─ u64_insert_intmap_checked             30.49 µs      │ 144.7 µs      │ 42.72 µs      │ 43.54 µs      │ 100     │ 100
├─ u64_insert_intmap_entry               42.44 µs      │ 164.1 µs      │ 43.67 µs      │ 45.6 µs       │ 100     │ 100
├─ u64_insert_intmap_without_capacity    482.3 µs      │ 612.3 µs      │ 500.5 µs      │ 523.8 µs      │ 100     │ 100
╰─ u64_resize_intmap                     31.01 µs      │ 36.46 µs      │ 31.06 µs      │ 31.31 µs      │ 100     │ 100
```

Closes #62