@jakoschiko 2024-09-01
Hmm, divan seems to be in conflict with the MSRV. Should I convert the benches to an own crate, similar to `integration_tests`?