@jakoschiko 2024-09-01
I tested [criterion](https://crates.io/crates/criterion). It would also be in conflict with MSRV.