@JesperAxelsson 2024-09-02
There is some more benchmarks in the `benchmark` branch. Getting it in a different crate might be the best options. It's not great having to switch branch just to test changes, but was an easy solution to not raise MSRV.