@JesperAxelsson 2024-09-02
Not sure if its possible to have different MSRV on dev and release mode. That would be a better solution otherwise.