@jakoschiko 2024-09-02
> Not sure if its possible to have different MSRV on dev and release mode. That would be a better solution otherwise.

We could change the CI like this:
- It uses the MSRV toolchain for compiling the lib
- It uses a newer Rust toolchain for compiling the lib and testing