@JesperAxelsson 2024-09-02
Yeah, might be the more practical solution. It should be easy to run benchmarks and test :)