@jakoschiko 2024-09-02
It still doesn't work, even though [it should work](https://doc.rust-lang.org/cargo/reference/specifying-dependencies.html#development-dependencies). There are a bunch of Rust issues about this topic. I'll investigate.