@jakoschiko 2024-09-02
Okay, as always, it's complicated.

- Even if it should work, it intentionally doesn't work, see https://github.com/rust-lang/cargo/issues/4988
- But after https://github.com/rust-lang/cargo/issues/5133, it works for `cargo install`! But we don't have anything to install. But it also works for `cargo build`! But only with a nightly flag...
- We could do this nice sed hack mentioned here https://github.com/rust-lang/cargo/issues/5133#issuecomment-1307094647
- Or we give up and move the benchmark into a separate crate...