@JesperAxelsson 2024-09-03
> Okay, moving it into a separate crate doesn't work either as long as it's part of the same workspace. I think I need to stop for now before I get an existential crisis.

Hehe, yeah I feel the pain. I can't understand why tooling is always so complicated.