@jakoschiko 2024-09-04
Ok, another attempt.

Now there are two completely unrelated folders:

1. The root of the repository
    - Contains the lib
    - Contains basic tests
    - Not a workspace anymore
1. The folder `integration_tests`
    - Contains more complicated tests (as before)
    - Contains the benchmarks (this is new)
    - Is now a workspace (this is new)

So if you want to test everything, you can do:

```sh
cargo test
cd integration_tests
cargo test
```

If you want to run all benchmarks, you can do:
```sh
cd integration_tests
cargo bench
```

Because we are using a separate workspace that is not related to the main lib, everything works. From now on tests and benchmarks can use new and shiny features without compromises.

Is this a feasible solution for you?

Can you also check if you're happy with the stability of the divan benchmarks? They run really fast, but are slightly less reliable (I think this is by design).