@JesperAxelsson 2024-09-07
The consistency seems better then before. I could some really strange measurements where one benchmark doubled in time between runs. It might be possible to increase the iteration count later if we feel the need. 

It does feel nice to have all the benchmarks and integration test separate. You can add packages without having to consider the main package.

Great job!